split site/tests into site/tests/lib and site/tests/routes
