+++ site/tests/routes/DiatomicPage.test.ts
@@ -1,7 +1,7 @@
+import Page from '$routes/tasks/diatomics/+page.svelte'
 import { mount } from 'svelte'
 import { beforeEach, describe, expect, test, vi } from 'vitest'
-import { doc_query } from '.'
-import Page from '../src/routes/tasks/diatomics/+page.svelte'
+import { doc_query } from '..'
 
 // Mock fetch function
 const mock_fetch = vi.fn()
