+++ site/tests/routes/ModelsPage.test.ts
@@ -1,6 +1,6 @@
+import Page from '$routes/models/+page.svelte'
 import { mount, tick } from 'svelte'
 import { beforeEach, describe, expect, it } from 'vitest'
-import Page from '../src/routes/models/+page.svelte'
 
 describe(`Models Page`, () => {
   beforeEach(() => {
